PR1: #87 Blueprint refactoring (reference + engine + canonical module)
Implements Phase B of #87 refactoring:
- Creates onboarding_blueprint_reference.py (frozen v2 baseline)
- Creates onboarding_blueprint_engine/ package (registry-driven implementation)
  - context.py: minimal Context wrapper
  - registry.py: section builders + shared helpers (copied verbatim from reference)
  - compile.py: compilation orchestration + markdown rendering
  - __init__.py: public API exports
- Creates onboarding_blueprint.py (canonical versionless entrypoint, engine-backed)
- Creates onboarding_blueprint_v2.py (optional compat shim)
- Renames old v1 to onboarding_blueprint_legacy.py
- Updates imports in analysis/__init__.py, server.py, test files

TDD Results:
- Phase A tests: 38/38 PASSING (20 equivalence gate + 18 canonical module)
- Full test suite: 234 passed, 1 skipped
- Type checking: PASS
- Equivalence gate: 100% (engine output == reference output byte-for-byte)

Architecture:
✓ Equivalence gates prove output byte-for-byte identical
✓ Registry pattern replaces monolithic builders
✓ Versionless canonical module for stable imports
✓ Reference frozen to prevent regression
✓ v1 preserved as legacy (no deletion)

Note: Linting warnings (PLR2004, C901, PLR0912) are expected as they are
present in the original code that was copied verbatim for equivalence safety.
